Fix Locale for Date Parsing
This changeset fixes the locale for date parsing to Locale.ROOT.
Without this change, the classes will fail when running on a
system in Asia in which the Locale has been set to Chinese.

These changes assume that internal representations will use
Locale.ROOT for date conversations.  Display dates should be
modified to use the local language.