output only path in css
This patch changes the css output to output only the path of the URI
rather than including the host.  This makes it consistent with
the javascript behavior, and is necessary for OpenGamma to work
properly if it is behind a proxy server, which can change the
name of the host.